Rename signature-discovery/client to intenral/signaturediscovery/client
