+++ launcher/internal/signaturediscovery/client_test.go
@@ -1,4 +1,4 @@
-package client
+package signaturediscovery
 
 import (
 	"context"
